example2/api/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
coloredlogs==15.0.1
datasets==2.14.5
dill==0.3.7
evaluate==0.4.1
exceptiongroup==1.1.3
fastapi==0.104.0
filelock==3.12.4
filetype==1.2.0
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.6.0
h11==0.14.0
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
install==1.3.5
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.2
numpy==1.26.1
onnx==1.14.1
onnxruntime==1.16.1
optimum==1.13.2
packaging==23.2
pandas==2.1.1
pdf2image==1.16.3
Pillow==10.1.0
pip==23.3
protobuf==4.24.4
pyarrow==13.0.0
pydantic==2.4.2
pydantic_core==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
responses==0.18.0
safetensors==0.4.0
sentencepiece==0.1.99
setuptools==68.0.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tokenizers==0.14.1
torch==2.1.0
tqdm==4.66.1
transformers==4.34.1
triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
uvicorn==0.23.2
wheel==0.41.2
xxhash==3.4.1
yarl==1.9.2
